Use ontoportal-client classes in BioPortal and AgroPortal implementations
These changes introduce an `OntoPortalInterface` base class. It expects subclasses to populate its `OntoPortalClientClass` and `api_key_name` class variables. `OntoPortalInterface` then handles all of the implementation details of constructing the `OntoPortalClient` instance and using it to make requests.
