+++ .gitignore
@@ -17,3 +17,4 @@ notebooks/api-key.txt
 
 .template.db
 .venv
+.tox/
